+++ api/V0ScaleDatastore.go
@@ -23,38 +23,22 @@ func (e *API) V0ScaleDatastore(w http.ResponseWriter, r *http.Request) {
 	d := e.ErrorMessage(0, "V0ScaleDatastore", "ok")
 
 	if vars["count"] != "" {
-		newCount, _ := strconv.Atoi(vars["count"])
-		oldCount := e.Config.DSMax
-		logrus.WithField("func", "api.V0ScaleDatastore").Debug("Scale current: ", oldCount)
-
-		e.Config.DSMax = newCount
-
-		d = []byte(strconv.Itoa(newCount - oldCount))
-
-		// Save current config
-		e.Redis.SaveConfig(*e.Config)
-
-		// if scale down, kill not needes agents
-		keys := e.Redis.GetAllRedisKeys(e.Framework.FrameworkName + ":datastore:*")
-
-		for keys.Next(e.Redis.CTX) {
-			key := e.Redis.GetRedisKey(keys.Val())
-			task := e.Mesos.DecodeTask(key)
-			task.Instances = newCount
-			e.Redis.SaveTaskRedis(task)
-
-			if newCount < oldCount {
-				e.Mesos.Kill(task.TaskID, task.Agent)
-				logrus.WithField("func", "api.V0ScaleDatastore").Debug("TaskID: ", task.TaskID)
-			}
-			if newCount > oldCount {
-				e.Mesos.Revive()
-			}
-			oldCount = oldCount - 1
+		count, err := strconv.Atoi(vars["count"])
+		if err != nil {
+			logrus.WithField("func", "api.V0ScaleDatastore").Error("Error: ", err.Error())
+			return
 		}
+		d = e.scaleDatastore(count)
 	}
 
 	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
 	w.Header().Set("Api-Service", "v0")
 	w.Write(d)
+}
+
+// scaleDatastore - can scale up and down the K8 datastore
+func (e *API) scaleDatastore(count int) []byte {
+	r := e.scale(count, e.Config.DSMax, ":datastore:*")
+	e.Config.DSMax = count
+	return r
 }